Move prettify_and_dedup_list test location
